Generalized the text to "interfaces"
The version of the text describes "interfaces" instead of merely "public APIs", as discussed in #90. I also upcased some RFC2119-words, and changed some text into the passive voice (not complete, only where it touched the text I was changing anyway). Tried not to fix line-wrap to ease diff'ing for humans.